third_party/WebKit/LayoutTests/http/tests/devtools/service-workers/service-workers-redundant-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
ServiceWorkers must be shown correctly even if there is a redundant worker.

The first ServiceWorker is activated.
==== ServiceWorkersView ====
127.0.0.1/devtools/service-workers/resources/service-worker-redundant-scope
Update
Unregister
Source
changing-worker.php
Received
Status
#N activated and is running
stop
Clients
Push
Push
Sync
Sync
============================
The second Serviceworker is installed.
==== ServiceWorkersView ====
127.0.0.1/devtools/service-workers/resources/service-worker-redundant-scope
Update
Unregister
Source
changing-worker.php
Received
Status
#N activated and is running
stop
#N waiting to activate
skipWaiting
Received
Clients
http://127.0.0.1:8000/devtools/resources/inspected-page.html
focus
Push
Push
Sync
Sync
============================
The first ServiceWorker worker became redundant and stopped.
==== ServiceWorkersView ====
127.0.0.1/devtools/service-workers/resources/service-worker-redundant-scope
Update
Unregister
Source
changing-worker.php
Received
Status
#N activated and is running
stop
Clients
Push
Push
Sync
Sync
============================
